// src/app/Shared/Models/FilterOption.model.ts
export class FilterOptions {
    filter?: string;
    sort?: string;
    search?: string;
    sortCode?: string;
    code?: string;
    sortName?: string;
    sortType?: string;
    statusSort?: string;
    userFeatureId?: string;
    companyId?: string;
    sortEmail?: string;
    ProjectId?: string;
    sortLabel?: string;
  }
  